Addings -/- Edits
1) - Player states machine;
2) - Animations for player;
3) - Animation changings;
4) - Combo function;
5) - Edited [Player] class;
6) - Damage exchange.

+++ src/include/main.hpp
@@ -3,7 +3,7 @@
 constexpr char _CopyrightString[] = "Copyright In Crucible 2022";
 constexpr char _NameOfApplicatoin[] = "From Grave to Heaven";
 
-long float _TimeAccumilator;
+float _TimeAccumilator;
 
 void GoToMainMenu   ();
 void UpdateMenu     ();
